+++ PSNovusTools/Private/Close-SSHSession.ps1
@@ -1,0 +1,13 @@
+function Close-SSHSession([SSH.SshSession]$session) {
+
+    try{
+        $result = Remove-SSHSession -SessionId $session.SessionId 
+      }
+      catch {
+          Write-Error "Failed to close SSH Session."
+      
+          
+      }
+
+      return $result   
+}