+++ Changelog.md
@@ -3,10 +3,17 @@
 <!-- TOC -->
 
 - [Changelog](#changelog)
+    - [Release v0.1.2](#release-v012)
     - [Release v0.1.1](#release-v011)
     - [Release v0.1.0](#release-v010)
 
 <!-- /TOC -->
+
+## Release v0.1.2
+
+* New function Copy-SSHFolder
+* Update Copy-SSHFile with new Close-SSHSession
+* New private Close-SSHSession function 
 
 
 ## Release v0.1.1